feat: consolidated files; add log level cache
