--- electron/common/logger/level/__tests__/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../../tsconfig.test.json"
-}
