--- electron/common/logger/level/get-log-level.ts
@@ -1,5 +1,0 @@
-import type { LogLevel } from '../types.js';
-
-export const getLogLevel = (): LogLevel => {
-  return (process.env.LOG_LEVEL || 'info') as LogLevel;
-};
