--- electron/common/string/includes-ignore-case.ts
@@ -1,9 +1,0 @@
-import type { Maybe } from '../types.js';
-import { equalsIgnoreCase } from './equals-ignore-case.js';
-
-export const includesIgnoreCase = (
-  values: Array<string>,
-  valueToFind: Maybe<string>
-): boolean => {
-  return values.some((value) => equalsIgnoreCase(value, valueToFind));
-};
