--- electron/common/string/to-upper-snake-case.ts
@@ -1,5 +1,0 @@
-import snakeCase from 'lodash-es/snakeCase.js';
-
-export const toUpperSnakeCase = (value: string): string => {
-  return snakeCase(value).toUpperCase();
-};
